Exercicio finalizado
Não consegui finalizar os desafios, apenas o primeiro
