--- semana 7/labenusers/src/components/HendleUsers.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-
-export default class HendleUsers extends React.Component{
-
-    render(){
-    
-    
-    return(
-  ""
-    
-    )
-    
-    
-    }
-}
-    
-    
-    
-
