--- semana 7/labenusers/src/components/Users.js
@@ -1,27 +1,0 @@
-import React from 'react'
-
-
-export default class Users extends React.Component{
-
-
-
-
-
-    render(){
-    
-    
-    return(
-
-""
-   
-    );
-    
-   
-    
-    }
-    
-    
-    
-    }
-    
-
