+++ semana7/labenusers/public/manifest.json
