+++ semana7/labenusers/src/parameters.js
@@ -1,0 +1,9 @@
+export const baseUrl = 'https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users'
+
+export const configHeader = {
+
+    headers: {
+      Authorization: "william-gomes-cruz"
+    }
+    
+  }
